requirements.txt
openai==0.26.5
discord.py==2.2.0
python-dotenv==1.0.0